+++ packages/gafl-webapp-service/src/handlers/__tests__/agreed-handler-recurring-payments.spec.js
@@ -5,8 +5,7 @@ import { preparePayment, prepareRecurringPaymentAgreement } from '../../processo
 import {
   sendPayment,
   sendRecurringPayment,
-  getPaymentStatus,
-  getRecurringPaymentAgreement
+  getPaymentStatus
 } from '../../services/payment/govuk-pay-service.js'
 import { prepareApiTransactionPayload } from '../../processors/api-transaction.js'
 import { v4 as uuidv4 } from 'uuid'
@@ -23,8 +22,7 @@ jest.mock('../../services/payment/govuk-pay-service.js', () => ({
     }
   })),
   getPaymentStatus: jest.fn(),
-  sendRecurringPayment: jest.fn(() => ({ agreementId: 'agr-eem-ent-id1' })),
-  getRecurringPaymentAgreement: jest.fn(() => ({ payment_instrument: { card_details: { last_digits_card_number: '1234' } } }))
+  sendRecurringPayment: jest.fn(() => ({ agreementId: 'agr-eem-ent-id1' }))
 }))
 jest.mock('../../processors/api-transaction.js')
 jest.mock('@defra-fish/connectors-lib')
@@ -110,14 +108,6 @@ describe('The agreed handler', () => {
       prepareRecurringPaymentAgreement.mockResolvedValueOnce(preparedPayment)
       await agreedHandler(getMockRequest(), getRequestToolkit())
       expect(sendRecurringPayment).toHaveBeenCalledWith(preparedPayment)
-    })
-
-    it('gets the recurring payment agreement information sending agreement id to Gov.UK Pay', async () => {
-      const agreementId = 'agreementId'
-      const mockSendRecurringPayment = { agreement_id: agreementId }
-      sendRecurringPayment.mockResolvedValueOnce(mockSendRecurringPayment)
-      await agreedHandler(getMockRequest(), getRequestToolkit())
-      expect(getRecurringPaymentAgreement).toHaveBeenCalledWith(agreementId)
     })
 
     describe('when there is a cost and recurringAgreement status is set to true', () => {
